Use cross-process safe locking
